Texier::Parser::Regexp now returns captured subgroups separately.
Previously, the result of regexp was allways single string - the whole matched string. Now, when the regexp contains groups, they are returned separately. 

For example, the expression e(/(hello) (world)/) when parsing string "hello world" previously returned ["hello world"] but now returns ["hello", "world"]. 

If there are no groups, the whole string is returned (as it was before).